Move const and b2api helpers into seperate modules
b2.console_tool module got very big making it hard to maintain.
This minimal refactor is done with preperation for refactoring of bash
autocomplete which would otherwise would have also had to be coded in
the 3k+ LoC console_tool.py .
