+++ b2/_cli/__init__.py
@@ -1,0 +1,9 @@
+######################################################################
+#
+# File: b2/_cli/__init__.py
+#
+# Copyright 2023 Backblaze Inc. All Rights Reserved.
+#
+# License https://www.backblaze.com/using_b2_code.html
+#
+######################################################################
